Split OriginAddress into OriginAuthority and OriginListeningAddress.
The former is used as an "authority" part of HTTP URLs. The latter --
for listening and connecting. URLs want something like "localhost" but
sockets cannot listen on "localhost".

Discovered that we cannot use default (i.e., undefined) listening host
because older nodejs (e.g., v0.10.25) essentially interpret it as "any
IPv4 if possible" while the newer nodejs interpret undefined host as "::
if possible". We now specify "::" explicitly in hope to avoid this
difference.  Here is one reference explaining the newer behavior:

https://github.com/nodejs/node-v0.x-archive/issues/9195#issuecomment-160754041