Simplified: Reduce nesting, avoid Mocha, and use recently added features.
Mocha is nice for dealing with a large number of small, fast,
isolated test cases [accumulated in one file]. However, when dealing
with a relatively long-running test that has multiple sub-cases, the
describe/it() architecture and inability to specify command line
parameters starts to get in the way.

I suspect this is especially true if one tries to run multiple such
tests in parallel, although that might be best done using parallel(1)
or similar external tools and not Javascript code (unless some
coordination among parallel tasks is needed).