Adds babel-cli and moves some dev dependencies to dependencies section.
Babel-cli is added to fix the `src/{client, proxy, server}/index.js`
files being callable. This only works if `node_modules/.bin` is
added to PATH (and current directory is the root of the project).

Other dependencies (babel dependencies and bluebird) are moved to enable
transpilation in a non-dev environment.