Moved Uri.{host,port} to Authority. Support IPv6 addresses.
Node.js uses several different anonymous address-like structures. Using
a named class may help identify and convert various "addresses":

* general-purpose "address" or "HostPort" structure: port + host
* [URI] authority: port + host (with IPv6 hosts bracketed when imaged)
* net.server.address(): port + address + family string
* options for socket.connect(): port + host + integer family + ...

At first, I added a more general class with 3+ fields (port, domain, and
ip) but it turned out to be a bad idea because most callers do not know
whether they have an IP address or a domain name, making the code
unnecessary complex. We may still need such a general "Address" class,
but we should not add it until it is needed, and it will probably not
belong to Uri.

XXX: It is easy to mistake one of the anonymous structures for Authority
or vice versa. Our code usually does not know which structure it deals
with (but it is usually HostPort). Converting all structures to
Authority is rather difficult because many are hard to find. These facts
make me question whether adding Authority was actually worth it.

Also support IPv6 addresses in URIs and (generated) Host headers.