Reduce HTTP status-code comparison bugs
Some code used strings to set and/or to test status-codes (e.g, "304").
Some code used numbers. Strict comparisons led to unexpected outcomes.
To spot such problems during development, we removed direct access to
the status-code field to provide safer, type-aware getters and setters.

I do not like the added API complexity/awkwardness. If a similar problem
affects another field, we may want to create some kind of a
NumberOrString class to encapsulate all of this.

I did not use ES6 getters and setters for new methods because they,
unlike, regular method calls, do not protect from typos in function
names.

Also allowed malformed non-digit status-codes when parsing responses. If
we store the code as a string, then there is no reason to be rigid when
parsing, especially in a do-as-told product.

Merged SQUID-498-integer-status-code (295200e) with a few fixes/changes.