Fixed distribution of --tests across test configurations
When --tests were configured explicitly, we allocated all of them to
testing the first configuration instead of testing each configuration
(to the extent possible).

Also interpret --concurrency-level as the number of required concurrent
tests for proper testing of a each configuration. If --tests is limited,
we still want to _properly_ test as many configurations as possible.