Support more than one from-server response
The API still focuses on the response to be sent during the first
transaction. It is not yet possible to configure or get a callback from
subsequent responses -- they simply mimic the first response.

The new functionality is usually triggered during unexpected cache
misses in a test case that hits the proxy with multiple requests (all
but one of which are expected to be hits). The old code would just
reject the connection. Such rejections led to somewhat confusing or
surprising error outcomes -- most testers do not expect to see a genuine
proxy error response in a hit/miss test.

The primary motivation for adding this feature was supporting tests that
continue running (in a more-or-less meaningful way) despite check
failures. Such testing is very useful for quick/dirty experiments where
the existing test code generates desirable traffic but the outcome of
specific checks is unimportant as long as that traffic is generated. For
example, I may want to know whether a non-collapsing DUT repeatedly
tries to cache responses when bombarded with many collapsable requests.

TODO: Add --keep-going support or equivalent.