Fixed cases that send a lot of data
When a transaction writes a lot of data to the node socket, the node
socket implementation buffers it all and then writes when it can. Our
checkpoint() method correctly delayed transaction end until all data has
been written (!this.socket.bufferSize), but there was no code to react
to the "drain" event when this.socket.bufferSize finally becomes 0.