Delay transaction's sentEverything() event until all is written
This is a follow up to commit a023c40 "Fixed cases that send a lot of
data". Known sentEverything() event users do not care whether the
to-be-sent data was internally produced. They only care whether it was
written to the network. (Ideally, they would prefer to know when it was
received by the other end, but we cannot measure that, of course.)