Made Message::forceChunked configurable via --chunk-bodies
... and renamed it to _chunkBody.

Setting forceChunked used to force Transfer-Encoding:chunked
even in messages without bodies (that is why it was called
"forceChunked" rather than "useChunked", I guess). I removed
that hack because it was in the way of a global/CLI option that
should be ignored for bodyless messages. I did not find any
current users of that hack. It can be restored by adding a
different/new configuration option.