Made chunk size configurable via --chunk-size
Also support withholding the last-chunk via --withhold-last-chunk. The
sender is unaware of this withholding because we want the corresponding
transaction to finish as if nothing happened (i.e. this is not a "pause
at last-chunk" feature).

TODO: All of this should be configurable for individual messages or even
encoders and not just globally.