Partially fixed concurrent tests without --tests
Daft could stop such tests too soon because it assumed that each test
has the same concurrency level. That assumption fails when the level is
not explicitly confiugured _and_ the config generators produce test
configurations with different concurrency levels.

This fix is partial because the code still does not compute the desired
number of tests correctly when concurrency changes from test to test.
That bug leads to wrong leftoverTests values (at least).