Initial server-side inter-test persistent connection support
The server Agent can now be configured to keep the accepted connection
after the corresponding transaction is over and then pass that
connection to another server Agent object (e.g., a server Agent from the
subsequent test case).

Also converted Agent::stop() API to use async/await instead of raw
promises. This conversion was probably not necessary, but I was
struggling with making the tests stop. I now know that the tests were
not stopping because server.closeAsync() waits for the server-accepted
pconn that was intentionally not closed. However, it was difficult to
discover that dependency, and raw promises made things worse.