Overlord: Support POP v4 and its shutdown-manner option
This support is useful for tests that check how the DUT reacts to
various restarting ways. For example, does a harsh kill-9 restart leaves
the disk cache in a consistent/usable state?

Also moved most of the Overlord request parameter formation code to
src/overlord/Command.js. This move started because a test wanted to
specify shutdown-manner, and it was difficult to pass that optional
information to ProxyOverlord without using risky option hashes and such.
Eventually, explicit passing of those specific options became
unnecessary because Command code added direct support for the
--dut-shutdown-manner command line option, but I left the extracted code
because it may become more useful if we need to pass other ProxyOverlord
parameters to commands.