Detail Proxy Overlord errors; reject truncated Overlord responses
Just knowing that Overlord failed (i.e. responded with non-200 status
code) is rarely enough, especially when Overlord logs are not easily
accessible in a CI environment.

Also detect and reject truncated Proxy Overlord responses. Such
responses have not been observed in the real world but they are easy to
detect and reject (instead of hoping that the JSON parser or our
JSON-consuming code will do the right thing when JSON is truncated).