Fix --response-ends-at-eof=true --body-size=<large> support
Caching tests with large EOF-terminated responses silently quit after
the server produced and started writing (in the background) a large
(e.g., 10MB) response body to the client.

I believe that happened because nobody could satisfy a
this.server.closeAsync() promise that server/Agent::_stop() is awaiting.
Normally, that promise is probably already satisfied at await call time
because there are no open server connections at that time. Only when a
server transaction needs time to write all 10MB of produced and buffered
data, the promise is unfulfilled at await time, with no code to fulfill
it.

Now the server transaction waits for the socket buffers to be fully
written and then calls checkpoint() which detects the end of the
no-longer-writing transaction and closes its connection socket,
satisfying awaited promiseToClose.

Also removed the possible implication that we might be reading after
calling socket.end(). By half-close definition, we can only be writing.