Expose Dut config() and DutConfig::cachingEnabled()
This is somewhat controversial because some code might interrogate the
DUT configuration prematurely (i.e. before it is finalized/settled).

This is done so that test cases can customize their behavior based on
the DUT configuration. In most cases, the required DUT configuration
aspect can be computed by the test code itself (because that code has
configured the DUT!), but that computation requires extra effort and may
be error prone (e.g., the logic of determining whether caching is
enabled is call order-sensitive and also depends on DUT defaults). In
theory, that DUT configuration might be even affected by the global Daft
configuration that a particular test code is not directly aware of.

----

Merged related SQUID-659-truncated-responses changes (as of 8778fb0).