Honor null server().[transaction.]response.body when sending a reply
... by not sending a response body at all. We used to generate and send
a default body instead, but we can only generate a body when the body is
undefined. When the body was explicitly configured, including configured
to be null, we should honor that configuration.

Fixing the above bug required improving AssertForwardedMessage(), our
forwarded message check, as well: When a Daft server sent an HTTP
response (without a Content-Length header) that was terminated by an EOF
without any body bytes (i.e. without a body from Daft server point of
view), Squid could forward that response as a response with a chunked
body containing nothing but last-chunk (i.e. an empty body from Daft
client point of view). This change would trigger false body forwarding
assertions.

----

Inspired by SQUID-659-truncated-responses changes (as of 8778fb0) that,
however, did not get this quite right and were sending a non-empty body
while attempting (but failing) to suppress that.