Remove support for concurrent HTTP transactions
... inside one Server instance.

Just close the new connection if the client manages to open another
connection to our origin server.

No known test needed concurrent transaction support, and removing that
support simplified the code. It is also possible that closed connections
would be easier to triage than concurrent responses, especially when
those concurrent responses happen to carry the same or similar IDs (or
share other characteristics that the test writer could expect to be
rather unique).

----

Based on SQUID-280-collapsed-slaves as of commit fde1cb7.