Better socket "drain" event description
While working on a new feature, I was surprised that the "drain" event
did not fire after "we wrote everything". The updated documentation
explains that this event has a precondition -- "we could not write
everything at once [earlier]". This precondition explains why this event
does not fire on every ordinary socket.write() call.