fixup: Moved request Range header field parsing to Header
I actually accidentally re-implemented the parser in Header because I
did not notice the existing one in Request. The new implementation may
be slightly more conservative/picky.

The new Header positions feels better because the original code did not
use any Request properties except for the this.header field itself.