Do not tell Server::Transaction what Range request to expect
Instead, generate a 206 response matching the actual Range request (if
any). This enhancement was added to test cases where the proxy may
forward a client Range request as a regular/full GET, but, in
retrospect, it was not really required for that because those cases
could predict whether the proxy would do the conversion. Still, it is an
improvement that may become essential in other cases.

The new request-aware Response finalization code needs polishing. It may
allow us to eventually address no default body in HEAD responses TODO,
even though we are still struggling with the exact meaning of this.body
presence.