Start upgrading [[low,high]...] ranges to proper classes
Also do not store parsed or configured ranges in Message, at least until
the code actually requires doing so. At the very least, this avoids
questions like "Where did this.ranges come from and what are we supposed
to do with them (that those who put them here could not)?". I suspect
that, in many cases, the higher-level code is in a better position to
store and manage these containers.

I am not quite sure where the [Ranges-related] parsing code belongs:

* It feels wrong to put it in http/Header and http/Response classes
  because doing so will eventually bloat those "base" classes with code
  specific to all the complex protocol parts we need to support.

* We could use Foo parsers as object Foo makers, and this code does that
  in some cases, but if parsing is complex, we may pollute those Foo
  classes as well, especially if complex parsing depends on HTTP
  version.

* IIRC, our earlier idea was to put HTTP version-sensitive code in
  src/http/one/.

This code needs more polishing.