Fix workerListeningAddresses()[0].port being 3130 instead of 3128
... when worker-dedicated listening ports were not enabled.

Workers never use port 3130, even when worker-dedicated ports are
enabled. The bug was not triggered until a new test wanted to send a
request to a worker without enabling worker-dedicated ports. Such
testing code is valid: The testing code may (and, in fact, often should)
be written to send requests to the address of a round-robin selected
worker, even if that address is always the same in _some_ test
configurations.