Revised Server::Agent::keepListening()
Official commit 6163502 added keepListening() to support multiple
sequential server transactions. The enhancement was required for Squid
cache_peer support and was used in squid-dafts::dead-peer test. In that
test, the first cache_peer transaction is a TCP probe that never sends a
request. AFAICT, the corresponding Server::Transaction never ends
because it waits to send an HTTP response (to the HTTP request that is
not coming on an already closed connection): "expecting to send more".
That bug prevents Server::Agent from stopping, which helped that test,
and the bug went unnoticed until now.

Branch commit 49b7523 added a keepListening() mode: never/once/always.
However, the key there, AFAICT, was not the addition of a third
(effectively unused!) mode, but the stop() method adjustments that
allowed the server to keep running after the first HTTP transaction was
over. 304 retrying tests could not rely on the bug mentioned above!

The needs of the old dead-peer and the new accumulate-headers-after-304
squid-daft tests can be met with just two keepListening() modes. Both
tests need the server to handle two connections. Neither test needs the
server to auto-stop after that. The tests could check that there were
exactly N server transactions, but that can be (and is probably best)
implemented as a check() rather than a listening limit.

Still, the official two-mode code was rather complex and probably
triggered some misunderstanding. I refactored it in this commit. The bug
mentioned in the first paragraph remains unconfirmed and unfixed.