Do not let Response::from(resource) overwrite default body bans
    <s HTTP/1.1 304 Not Modified\r\n
    ...
    will send the entire response body [15 bytes]

The Response::from(resource) method is meant to inform the response what
resource it is based on. It should not _force_ the presence of a
response body (and probably any other specific response part!) because a
response is just a (partial) "view" of that resource. By default,
request method and configured response status code (e.g., 304) determine
response body presence in a transaction.

Some test cases assume that calling Response::from(resource) with a
body-carrying resource forces a body to be sent, despite the response
status code (e.g., 101 Switching Protocols). These test cases must be
adjusted to use the new Body::forcePresence() API.

We continue to suffer from the poorly defined Body class scope. It is
likely that the class should be split into several ones (e.g., one for
sending a body and one for receiving a body). See also: Commit b07fc6e.

This fix took several iterations to complete. The following ideas were
also implemented but subsequently rejected:

* Save a Resource::body clone but postpone creation of Response::body
  until response finalization. Some existing tests assume that calling
  Response::from(resource) with a body-carrying resource makes
  Response::body not nil. These test cases would fail when dereferencing
  that still-nil body data member. The correct long-term solution here
  is probably to _always_ create Response::body, just like we always
  create Response::header (see the next bullet for tests that need to
  disable response body sending). Going in the opposite direction from
  that long-term solution felt wrong.

* Adding Body::forceAbsence() API for tests that need to disable
  response body sending. Unlike the current `body = null` approach,
  using that API would not trigger nil pointer dereference errors. That
  API was not added (and the migration away from `body = null` was not
  started) because there is a suspicion that those dereference errors
  indicate code bugs. For example, configuration code that blindly sets
  (absent) body properties is likely to be conceptually broken OR should
  be made unreachable. We will wait for a strong "pro" use case before
  adding Body::forceAbsence() and migrating away from `body = null`.