Preserve perceived addition order of check() and addMiss/HitCheck()
When a test calls, for example, testCase.check() followed by a call to
testCase.addMissCheck(), we expect the checks to be executed in the same
order. Before this change, the order was reversed because addMissCheck()
and addHitCheck() added checks to each test case _client_, and client
checks run _before_ the case-wide checks. Besides being a surprise,
reversed order may complicate deciphering check failures.