fixup: Import order polishing and diff reduction
There does not seem to be One True Way for ordering import statements.
Until we adopt some fancy plugin that will do it for us, we want to sort
them (as case-insensitive lines) with an editor.

The only manual action now is the separation of our modules (first) from
3rd-party modules. I am not sure that separation is a good idea, but it
feels like these two sets of modules may be treated somewhat differently
in some cases.