Revised Server::Agent::keepListening()
Official commit 6163502 added keepListening() to support multiple
sequential server transactions. The enhancement was required for Squid
cache_peer support and was used in squid-dafts::dead-peer test. In that
test, the first cache_peer transaction is a TCP probe that never sends a
request. AFAICT, the corresponding Server::Transaction never ends
because it waits to send an HTTP response (to the HTTP request that is
not coming on an already closed connection): "expecting to send more".
That bug prevented Server::Agent from stopping, which helped that test,
and the bug went unnoticed until accumulate-headers-after-304 (some of
its early versions were probably not compatible with that bug).

The needs of the old dead-peer and the new accumulate-headers-after-304
squid-daft tests could have been met without serious refactoring: Both
tests need the server to handle two connections. Neither test needs the
server to auto-stop after that. The tests could check that there were
exactly N server transactions, but that can be (and is probably best)
implemented as a check() rather than a listening limit.

Still, the original code was rather complex and probably triggered some
costly misunderstanding. I refactored it in this commit. The bug
mentioned in the first paragraph remains unconfirmed and unfixed.

----

Based on SQUID-887-304-headers-accumulation commits 813c5c9, f223e4d,
78f6e30, bec3007, and 49b7523.