Allow Transaction users to customize checkpoint()
Such customizations are useful for triggering various complex events
that would be awkward to support via a regular blockUntilFoo() or
didFoo() methods. For example, "foo" can be "received 13 body bytes".

On the other hand, if we discover adding the same custom checkpoint code
to multiple test cases/transactions, then that code becomes a good
candidate for conversion into a regular Transaction method.