Support --dut-memory-cache-size
Smaller caches are handy for tests that must fill most of that cache
space. Filling 100+ MB of cache space is doable, but there is no reason
to trigger so much noise/overhead.

Larger caches are still a good default for caching tests that do not
want to worry about running out of cache space (or metadata "inodes").

The added XXX is about an old/existing problem with command line options
that should be solved using a dedicated project.